requirements.txt
opencv-python
streamlit
matplotlib
plotly
numpy
pandas
scikit-learn
python-dotenv
tensorflow
keras
deepface
seaborn
pillow